#1 Initialize SnackBarManager in module level for simplicity
